Module Caqti_driver_pgx

PostgreSQL driver for Caqti based on pgx

This driver is implemented in terms of the pgx library.

It handles URIs of the form

pgx://<user>:<password>@<host-or-directory>:<port>/<database>

where <host-or-directory> is either a host name, an IP number, or a directory containing the Unix domain socket of a locally running PostgreSQL server. In the latter case, the directory must be percent-encoded, e.g.

pgx://jdoe@%2fvar%2frun%2fpostgresql

The interface provided by this module should normally not be used by applications, but provides access to some pgx specifics in case they are needed.

Error Details

type Caqti_error.msg +=

| Pgx_msg of string * Pgx.Error_response.t